+++ pikoci/transport/http/route_names.go
@@ -34,12 +34,15 @@ const (
 	GetPipelineJob
 
 	CreateJobBuild
+	CreateRetryJobBuild
 	UpdateJobBuild
 	DeleteJobBuild
 	ListJobBuilds
 	InsertBuildGetVersion
+	FindBuildGetVersions
 	GetJobBuild
 	CancelJobBuild
+	RetryJobBuild
 
 	GetPipelineResource
 	UpdatePipelineResource